start of refactoring, WAHM
